编辑模块：将缓冲区偏移量和大小数据类型改为整型，并修复比特位类型数据偏移量计算问题

+++ src/app/module/edit/details/DetailTable.h
@@ -156,8 +156,8 @@ private:
     void setRowData(int row, const Icd::VehiclePtr &vehicle);
     void setRowData(int row, const Icd::SystemPtr &system);
     void setRowData(int row, const Icd::TablePtr &table);
-    void setRowData(int row, const Icd::ItemPtr &item, double offset);
-    void setRowOffsetAndSize(int row, const Icd::ItemPtr &item, double offset);
+    void setRowData(int row, const Icd::ItemPtr &item, int offset);
+    void setRowOffsetAndSize(int row, const Icd::ItemPtr &item, int offset);
 
     int insertRow(int row, const Icd::ObjectPtr &object);
