+++ src/core/icdcore/icd_item.h
@@ -45,9 +45,9 @@ public:
 
     int itemOffset() const;
     char *buffer() const;
-    virtual double bufferSize() const;
-    virtual double bufferOffset() const;
-    virtual double localOffset() const;
+    virtual int bufferSize() const;
+    virtual int bufferOffset() const;
+    virtual int localOffset() const;
 
     virtual double data() const { return 0.0; }
     virtual void setData(double d) { (void)d; }
@@ -86,8 +86,8 @@ protected:
 
     virtual void setBuffer(char *buffer);
     virtual void setItemOffset(int offset);
-    virtual void setBufferSize(double size);
-    virtual void setBufferOffset(double offset);
+    virtual void setBufferSize(int size);
+    virtual void setBufferOffset(int offset);
     virtual void adjustBufferOffset();
 
     // Serializable interface
